# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CATAMS API
  description: |
    Casual Academic Time Allocation Management System (CATAMS) API specification.
    
    This API provides endpoints for managing academic staff time allocation, including:
    - User authentication and management
    - Timesheet creation and management
    - Approval workflow management
    
    The system supports three main roles:
    - **LECTURER**: Can create and approve timesheets for tutors
    - **TUTOR**: Can view their timesheets and request modifications
    - **ADMIN**: Has full system access including HR functions
  version: 1.0.0
  contact:
    name: CATAMS Development Team
    email: dev@catams.edu.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.catams.edu.au
    description: Production server
...
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Create new user
      description: |
        Create a new user account in the system.
        
        **Required role**: ADMIN
        
        **Business rules**:
        - Email must be unique in the system
        - Password must meet security requirements (8+ chars, mixed case, numbers, special chars)
        - Only ADMIN users can create new accounts
        - Default user status is active
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              create_lecturer:
                summary: Create lecturer account
                value:
                  email: "new.lecturer@university.edu.au"
                  name: "Dr. Jane Doe"
                  password: "SecurePass123!"
                  role: "LECTURER"
              create_tutor:
                summary: Create tutor account
                value:
                  email: "new.tutor@university.edu.au"
                  name: "Alice Johnson"
                  password: "MyPassword456!"
                  role: "TUTOR"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              examples:
                created_user:
                  summary: Successfully created user
                  value:
                    id: 123
                    email: "new.lecturer@university.edu.au"
                    name: "Dr. Jane Doe"
                    role: "LECTURER"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                email_exists:
                  summary: Email already exists
                  value:
                    timestamp: "2025-08-01T10:15:30.123Z"
                    status: 400
                    error: "EMAIL_EXISTS"
                    message: "Email already exists"
                    path: "/api/users"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    UserCreateRequest:
      type: object
      required:
        - email
        - name
        - password
        - role
      properties:
        email:
          type: string
          format: email
          description: User email address (must be unique)
          example: "new.lecturer@university.edu.au"
          maxLength: 255
        name:
          type: string
          description: Full name of the user
          example: "Dr. Jane Doe"
          maxLength: 100
        password:
          type: string
          format: password
          description: Initial password for the user
          example: "SecurePass123!"
          minLength: 8
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
    UserResponse:
      type: object
      required:
        - id
        - email
        - name
        - role
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
          example: 123
        email:
          type: string
          format: email
          description: User email address
          example: "lecturer@university.edu.au"
        name:
          type: string
          description: Full name of the user
          example: "Dr. John Smith"
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum:
        - ADMIN
        - LECTURER
        - TUTOR
      description: |
        User role in the system:
        - **ADMIN**: Full system access, HR functions
        - **LECTURER**: Can manage timesheets for their courses
        - **TUTOR**: Can view their own timesheets and request modifications
      example: "LECTURER"

    # Timesheet Management Schemas
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
          example: "2025-08-01T10:15:30.123Z"
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_FAILED"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed: email: Email format is invalid"
        path:
          type: string
          description: Request path that caused the error
          example: "/api/auth/login"

# Security configuration
security:
  - bearerAuth: []